Use executeExchange for test API
