+++ package.json
@@ -13,11 +13,7 @@
     "lint": "eslint \"src/**/*.ts*\"",
     "release": "np",
     "start": "vite --open --mode ssr",
-    "qwik": "qwik",
-    "api": "node ./api/main.js"
-  },
-  "stackblitz": {
-    "startCommand": "yarn dev & yarn api"
+    "qwik": "qwik"
   },
   "main": "./lib/index.qwik.cjs",
   "qwik": "./lib/index.qwik.mjs",
@@ -67,13 +63,11 @@
     "@typescript-eslint/eslint-plugin": "5.40.1",
     "@typescript-eslint/parser": "^5.40.1",
     "@urql/exchange-auth": "^1.0.0",
+    "@urql/exchange-execute": "^2.0.0",
     "@urql/exchange-graphcache": "^5.0.2",
     "cookie": "^0.5.0",
-    "cors": "^2.8.5",
     "eslint": "8.25.0",
     "eslint-plugin-qwik": "latest",
-    "express": "^4.18.2",
-    "express-graphql": "^0.12.0",
     "graphql": "^16.6.0",
     "node-fetch": "3.2.10",
     "np": "7.6.2",
